requirements.txt
pytest==8.4.1
selenium==4.34.2
pyYAML==6.0.2
mysql-connector-python==9.4.0
pyodbc==5.2.0
python-docx==1.2.0
pytest-xdist==3.8.0
openpyxl==3.1.5
pywinauto==0.6.9
Faker~=16.7.0
pyautogui
opencv-python 
allure-pytest